Writers now ensure they do not overflow their RingBuffer
This is done in the Writer instead of the RingBuffer, in the spirit of
the LMAX disruptor talk, to keep the RingBuffer simple and allow us to
enact differing policies in our Writers.

We now have to pass in the the number of writers going into the
RingBuffer and the Readers of the RingBuffer, too, in order to make
sure that the Readers have kept up with where the Writers are.

This numOfWriters variable should probably be removed and only added
in another "MultiWriter" class that allows more than one writer to a
buffer, mayhaps.